# apps/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MissionHeartbeatRequest": {
        "description": "任務ハートビートリクエスト。",
        "properties": {
          "deviceId": {
            "title": "Deviceid",
            "type": "string"
          }
        },
        "required": [
          "deviceId"
        ],
        "title": "MissionHeartbeatRequest",
        "type": "object"
      },
...
      "ReportFormatCreate": {
        "description": "レポート種別の新規作成。",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ReportFormatCreate",
        "type": "object"
      },
...
      "ReportFormatTemplateItem": {
        "description": "種別に紐づくテンプレート 1 件（構成の追加・一括更新用）。",
        "properties": {
          "sortOrder": {
            "minimum": 0.0,
            "title": "Sortorder",
            "type": "integer"
          },
          "templateId": {
            "format": "uuid",
            "title": "Templateid",
            "type": "string"
          }
        },
        "required": [
          "templateId",
          "sortOrder"
        ],
        "title": "ReportFormatTemplateItem",
        "type": "object"
      },
      "ReportFormatTemplatesUpdate": {
        "description": "種別のテンプレート構成を一括更新。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportFormatTemplateItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "ReportFormatTemplatesUpdate",
        "type": "object"
      },
      "ReportFormatUpdate": {
        "description": "レポート種別の更新。",
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "ReportFormatUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "現場報告管理システム Citadel のバックエンド API",
    "title": "Citadel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/missions": {
      "get": {
        "description": "派遣中の任務一覧（管理画面用）。status 指定時はその状態の任務のみ。",
        "operationId": "list_missions_api_missions_get",
        "parameters": [
          {
            "description": "Active / Expired / Purged / Returned",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Active / Expired / Purged / Returned",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Missions",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{mission_id}/heartbeat": {
      "post": {
        "description": "任務のハートビート。Scout が「生存している」と報告する。\n利用停止/期限切れの場合は 403 と code: PURGED を返す。",
        "operationId": "mission_heartbeat_api_missions__mission_id__heartbeat_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionHeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mission Heartbeat",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{mission_id}/purge": {
      "post": {
        "description": "任務を利用停止（Purge）する。派遣名簿から削除し、該当 Scout は次回同期で 403 を受ける。",
        "operationId": "mission_purge_api_missions__mission_id__purge_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mission Purge",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{mission_id}/status": {
      "get": {
        "description": "任務の状態を取得（Admin のファイルインポート時など）。",
        "operationId": "mission_status_api_missions__mission_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mission Status",
        "tags": [
          "missions"
        ]
      }
    },
...
    "/api/report-formats": {
      "get": {
        "description": "レポート種別一覧を返す。Scout の報告書種別ドロップダウン等で使用。",
        "operationId": "list_report_formats_api_report_formats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Report Formats",
        "tags": [
          "report-formats"
        ]
      },
      "post": {
        "description": "レポート種別を 1 件作成。id は UUID で自動生成。",
        "operationId": "create_report_format_api_report_formats_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportFormatCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Report Format",
        "tags": [
          "report-formats"
        ]
      }
    },
    "/api/report-formats/{format_id}": {
      "delete": {
        "description": "レポート種別とそのテンプレート構成を削除。",
        "operationId": "delete_report_format_api_report_formats__format_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "format_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Format Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Report Format",
        "tags": [
          "report-formats"
        ]
      },
      "put": {
        "description": "レポート種別の名前を更新。",
        "operationId": "update_report_format_api_report_formats__format_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "format_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Format Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportFormatUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Report Format",
        "tags": [
          "report-formats"
        ]
      }
    },
    "/api/report-formats/{format_id}/templates": {
      "get": {
        "description": "指定したレポート種別に紐づくテンプレート部品を sort_order 順で返す。",
        "operationId": "get_format_templates_api_report_formats__format_id__templates_get",
        "parameters": [
          {
            "in": "path",
            "name": "format_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Format Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Format Templates",
        "tags": [
          "report-formats"
        ]
      },
      "put": {
        "description": "指定したレポート種別のテンプレート構成を一括更新。既存の紐づけは削除し、body で指定した並びで再登録。",
        "operationId": "update_format_templates_api_report_formats__format_id__templates_put",
        "parameters": [
          {
            "in": "path",
            "name": "format_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Format Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportFormatTemplatesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Format Templates",
        "tags": [
          "report-formats"
        ]
      }
    },
...
    "/api/reports/{report_id}": {
      "delete": {
        "description": "指定IDのレポートを削除。関連データを子から親の順でカスケード削除する。",
        "operationId": "delete_report_api_reports__report_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Report",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}